src/data_app/pyproject.toml
[tool.poetry]
name = "services"
version = "0.1.0"
description = ""
authors = ["vict0rcarvalh0 <victordecarvalho342@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
clickhouse-connect = "^0.7.19"
minio = "^7.2.8"
flask = "^3.0.3"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
boto3 = "^1.35.5"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
elasticsearch = "^8.15.0"
chardet = "^5.2.0"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
croniter = "^3.0.3"
scikit-learn = "^1.5.1"
plotly = "^5.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
